BACK-2777: Move commands to lib dir
