+++ lib/commands/config.js
@@ -14,24 +14,31 @@
  */
 
 const async = require('async');
-const init = require('../lib/init.js');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
-const handleActionFailure = require('./../lib/util').handleCommandFailure;
+const init = require('../init.js');
+const project = require('../project.js');
+const user = require('../user.js');
+const util = require('../util.js');
+const handleActionFailure = require('../util').handleCommandFailure;
 
-function list(argv, cb) {
+function initUrl(host, cb) {
+  if (host != null) user.host = util.formatHost(host);
+  cb();
+}
+
+function configure(argv, cb) {
   init(argv);
   return async.series([
+    (next) => initUrl(argv.instance, next),
     (next) => user.setup(argv, next),
-    (next) => project.restore(next),
-    (next) => project.list(next)
+    (next) => project.config(argv, next),
+    (next) => user.save(next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'list',
-  desc: 'list Internal Flex Services for the current app',
-  handler: list
+  command: 'config [instance]',
+  desc: "set project options (including optional Kinvey instance, e.g. 'acme-us1')",
+  handler: configure
 };
