+++ lib/commands/deploy.js
@@ -14,31 +14,27 @@
  */
 
 const async = require('async');
-const init = require('../lib/init.js');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
-const util = require('../lib/util.js');
-const handleActionFailure = require('./../lib/util').handleCommandFailure;
+const config = require('config');
+const service = require('../service.js');
+const init = require('../init.js');
+const project = require('../project.js');
+const user = require('../user.js');
+const handleActionFailure = require('../util').handleCommandFailure;
 
-function initUrl(host, cb) {
-  if (host != null) user.host = util.formatHost(host);
-  cb();
-}
-
-function configure(argv, cb) {
+function deploy(argv, cb) {
   init(argv);
-  return async.series([
-    (next) => initUrl(argv.instance, next),
+  return async.waterfall([
     (next) => user.setup(argv, next),
-    (next) => project.config(argv, next),
-    (next) => user.save(next)
+    (next) => project.restore(next),
+    (next) => service.validate(config.paths.package, next),
+    (version, next) => service.deploy(config.paths.package, version, next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'config [instance]',
-  desc: "set project options (including optional Kinvey instance, e.g. 'acme-us1')",
-  handler: configure
+  command: 'deploy',
+  desc: 'deploy the current project to the Kinvey FlexService Runtime',
+  handler: deploy
 };
