+++ lib/commands/list.js
@@ -14,21 +14,24 @@
  */
 
 const async = require('async');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
-const handleActionFailure = require('./../lib/util').handleCommandFailure;
+const init = require('../init.js');
+const project = require('../project.js');
+const user = require('../user.js');
+const handleActionFailure = require('../util').handleCommandFailure;
 
-function logout(argv, cb) {
+function list(argv, cb) {
+  init(argv);
   return async.series([
-    (next) => user.logout(next),
-    (next) => project.logout(next)
+    (next) => user.setup(argv, next),
+    (next) => project.restore(next),
+    (next) => project.list(next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'logout',
-  desc: 'resets host, removes sessions, and clears project settings',
-  handler: logout
+  command: 'list',
+  desc: 'list Internal Flex Services for the current app',
+  handler: list
 };
