+++ lib/commands/logout.js
@@ -14,27 +14,21 @@
  */
 
 const async = require('async');
-const config = require('config');
-const service = require('../lib/service.js');
-const init = require('../lib/init.js');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
-const handleActionFailure = require('./../lib/util').handleCommandFailure;
+const project = require('../project.js');
+const user = require('../user.js');
+const handleActionFailure = require('../util').handleCommandFailure;
 
-function deploy(argv, cb) {
-  init(argv);
-  return async.waterfall([
-    (next) => user.setup(argv, next),
-    (next) => project.restore(next),
-    (next) => service.validate(config.paths.package, next),
-    (version, next) => service.deploy(config.paths.package, version, next)
+function logout(argv, cb) {
+  return async.series([
+    (next) => user.logout(next),
+    (next) => project.logout(next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'deploy',
-  desc: 'deploy the current project to the Kinvey FlexService Runtime',
-  handler: deploy
+  command: 'logout',
+  desc: 'resets host, removes sessions, and clears project settings',
+  handler: logout
 };
