+++ lib/commands/logs.js
@@ -18,13 +18,13 @@ const chalk = require('chalk');
 const moment = require('moment');
 
 const config = require('config');
-const KinveyError = require('../lib/kinvey-error');
-const service = require('../lib/service.js');
-const init = require('../lib/init.js');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
-const handleActionFailure = require('./../lib/util').handleCommandFailure;
-const LogErrorMessages = require('../lib/constants').LogErrorMessages;
+const KinveyError = require('../kinvey-error');
+const service = require('../service.js');
+const init = require('../init.js');
+const project = require('../project.js');
+const user = require('../user.js');
+const handleActionFailure = require('../util').handleCommandFailure;
+const LogErrorMessages = require('../constants').LogErrorMessages;
 
 function isValidTimestamp(ts) {
   if (ts == null) return true;
