+++ lib/commands/recycle.js
@@ -14,26 +14,25 @@
  */
 
 const async = require('async');
-const service = require('../lib/service.js');
-const init = require('../lib/init.js');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
-const handleActionFailure = require('./../lib/util').handleCommandFailure;
+const service = require('../service.js');
+const init = require('../init.js');
+const project = require('../project.js');
+const user = require('../user.js');
+const handleActionFailure = require('../util').handleCommandFailure;
 
-function status(argv, cb) {
+function recycle(argv, cb) {
   init(argv);
   return async.series([
     (next) => user.setup(argv, next),
     (next) => project.restore(next),
-    (next) => service.jobStatus(argv.id, next)
+    (next) => service.recycle(next)
   ], (err) => {
-    // TODO: cb should receive result from tasks
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'job [id]',
-  desc: 'return the job status of a deploy command',
-  handler: status
+  command: 'recycle',
+  desc: 'recycle the Service',
+  handler: recycle
 };
