+++ lib/commands/status.js
@@ -14,11 +14,11 @@
  */
 
 const async = require('async');
-const service = require('../lib/service.js');
-const init = require('../lib/init.js');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
-const handleActionFailure = require('./../lib/util').handleCommandFailure;
+const service = require('../service.js');
+const init = require('../init.js');
+const project = require('../project.js');
+const user = require('../user.js');
+const handleActionFailure = require('../util').handleCommandFailure;
 
 function status(argv, cb) {
   init(argv);
