+++ test/tests-helper.js
@@ -160,7 +160,7 @@ helper.setup = {
     mockServer.apps();
     mockServer.dataLinks();
 
-    require('./../cmd/config').handler({}, (err) => {
+    require('../lib/commands/config').handler({}, (err) => {
       expect(err).to.not.exist;
       expect(mockServer.isDone()).to.be.true;
 
@@ -204,7 +204,7 @@ helper.setup = {
   initiateJobDeploy(mockServer, cb) {
     mockServer.deployJob();
 
-    require('./../cmd/deploy').handler({}, (err) => {
+    require('../lib/commands/deploy').handler({}, (err) => {
       expect(err).to.not.exist;
       expect(mockServer.isDone()).to.be.true;
 
@@ -237,12 +237,12 @@ helper.setup = {
   // Ensure modules are reloaded every time and tests are independent (e.g class User -> this.token will be cleared).
   clearRequireCache() {
     const modules = [
-      '/cmd/config', '/cmd/deploy', '/cmd/job', '/cmd/list', '/cmd/logout', '/cmd/logs', '/cmd/recycle', '/cmd/status',
-      '/lib/project', '/lib/service', '/lib/user', '/lib/util'
+      '/commands/config', '/commands/deploy', '/commands/job', '/commands/list', '/commands/logout', '/commands/logs', '/commands/recycle', '/commands/status',
+      '/project', '/service', '/user', '/util'
     ];
 
     modules.forEach(module => {
-      const pathToResolve = `./..${module}`;
+      const pathToResolve = `./../lib${module}`;
       delete require.cache[require.resolve(pathToResolve)];
     });
   },
